refactor: extract soft-apply and pending-local-override controllers
Split sync-controller.ts (1518 → 1138 lines, -25%) into two focused
modules:

- soft-apply-controller.ts: owns the soft-apply state machine (active
  session, timer, cooldown, programmatic-apply wiring) and all related
  suppress/cancel decisions
- pending-local-override.ts: owns the pending seek/ratechange guard
  lifecycle (remember, resolve, clear)

Both follow the existing createRoomStateApplyController factory pattern.
sync-controller.ts is now an orchestration layer that wires the modules
together. No behaviour changes; all 170 tests continue to pass.

Closes #38
